`GotImageThumbnail` moved to its own file
