+++ MVVM-C/Controllers/Article Details/DetailViewVC.swift
@@ -110,30 +110,3 @@ final class DetailViewVC: UIViewController {
         ])
     }
 }
-
-
-class GoTImageThumbnail: UIImageView {
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        configure()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    private func configure() {
-        layer.cornerRadius = 20
-        clipsToBounds = true
-        translatesAutoresizingMaskIntoConstraints = false
-    }
-    
-    func downloadThumbnail(_ link: String) {
-        NetworkManager.shared.downloadImage(from: link) { [weak self] (image) in
-            DispatchQueue.main.async {
-                self?.image = image
-            }
-        }
-    }
-}
